requirements.txt
pygame-ce==2.5.6
PyTMX==3.32
pyscroll==2.31